fix bug

+++ MainWindow.xaml.cs
@@ -33,8 +33,8 @@ namespace GTAIVSetupUtilityWPF
 
         public MainWindow()
         {
-            File.Delete("GTAIVSetupUtilityLogOld.txt");
-            File.Move("GTAIVSetupUtilityLog.txt", "GTAIVSetupUtilityLogOld.txt");
+            if ( File.Exists("GTAIVSetupUtilityLogOld.txt")) { File.Delete("GTAIVSetupUtilityLogOld.txt"); }
+            if ( File.Exists("GTAIVSetupUtilityLog.txt")) { File.Move("GTAIVSetupUtilityLog.txt", "GTAIVSetupUtilityLogOld.txt"); }
             NLog.LogManager.Setup().LoadConfiguration(builder =>
             {
                 builder.ForLogger().FilterMinLevel(LogLevel.Debug).WriteToFile(fileName: "GTAIVSetupUtilityLog.txt");
